Change type of chooseRow return
Changed the return value of chooseRow return so that it does not
typecast to int but stays numpy.int64

+++ test/python/transpiler/test_perm_row_col_synthesis.py
@@ -43,7 +43,7 @@ class TestPermRowColSynthesis(QiskitTestCase):
 
         instance = synthesis.choose_row(vertices, parity_mat)
 
-        self.assertIsInstance(instance, int)
+        self.assertIsInstance(instance, np.int64)
 
     def test_choose_row_returns_correct_index(self):
         """
